Gtk3: Fix warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38147 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/gtk3/widgets/drivemodelwidget.c
@@ -129,7 +129,7 @@ GtkWidget *drive_model_widget_create(int unit)
     GSList *group = NULL;
     size_t i;
     int type;
-    int num;
+    size_t num;
     int row;
 
     resources_get_int_sprintf("Drive%dType", &type, unit);
@@ -158,7 +158,7 @@ GtkWidget *drive_model_widget_create(int unit)
         g_signal_connect(radio, "toggled", G_CALLBACK(on_radio_toggled),
                 GINT_TO_POINTER(list[i].id));
 
-        gtk_grid_attach(GTK_GRID(grid), radio, 0, i + 1, 1, 1);
+        gtk_grid_attach(GTK_GRID(grid), radio, 0, (gint)(i + 1), 1, 1);
         last = GTK_RADIO_BUTTON(radio);
     }
 
@@ -203,7 +203,7 @@ void drive_model_widget_update(GtkWidget *widget)
     size_t i;
     int unit;
     int type;
-    int num;
+    size_t num;
     int row;
 
     unit = GPOINTER_TO_INT(g_object_get_data(G_OBJECT(widget), "UnitNumber"));
@@ -219,7 +219,8 @@ void drive_model_widget_update(GtkWidget *widget)
 
     for (i = 0; list[i].name != NULL && i < num / 2; i++) {
 
-        GtkWidget *radio = gtk_grid_get_child_at(GTK_GRID(widget), 0, i + 1);
+        GtkWidget *radio = gtk_grid_get_child_at(GTK_GRID(widget),
+                0, (gint)(i + 1));
         if (radio != NULL && GTK_IS_RADIO_BUTTON(radio)) {
             gtk_widget_set_sensitive(radio,
                     drive_check_type((unsigned int)(list[i].id),
